oe-selftest: devtool: avoid parallel races by using temporary copy of meta
Some of the devtool tests make changes to files under meta/ -
legitimately since we want these tests to be working with real recipes
and associated files. Unfortunately with the new oe-selftest
parallelisation this can break other tests if files go missing at the
wrong time (among other scenarios). To avoid this issue, simply take a
copy of the meta/ directory and use that for these tests.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>